refactor(a11y): relocate goal chart summary utils to chart type directory
